Major Update Progress
Code refactoring and update for 0.3.0 launch.
